[Modify] Modify conditions to hide dummy data and change the history link component to Timeline.
